+++ src/redux/intakeDashboard/reducers.js
@@ -1,0 +1,38 @@
+import { constants } from 'utils/constants';
+import actions from './actions';
+
+const initialState = {
+  items: {},
+  isLoading: false,
+  total: 0,
+  offset: 0,
+  error: null,
+  isCreating: false,
+};
+
+export default function intakeDashboardReducer(state = initialState, action) {
+  switch (action.type) {
+    case actions.FETCH_DAILY_INTAKE_COUNTS_REQUEST: {
+      return {
+        ...state,
+        isLoading: true,
+      };
+    }
+    case actions.FETCH_DAILY_INTAKE_COUNTS_SUCCESS: {
+      return {
+        ...state,
+        isLoading: false,
+        items: action.payload.data.result,
+      };
+    }
+    case actions.FETCH_DAILY_INTAKE_COUNTS_FAILURE: {
+      return {
+        ...state,
+        isLoading: false,
+        error: action.payload.data,
+      };
+    }
+    default:
+      return state;
+  }
+}